@@ -3,12 +3,15 @@
 source 'https://rubygems.org'
 git_source(:github) { |repo| "https://github.com/#{repo}.git"}
 
-gem 'rails', '~> 6.1.0'
+gem 'rails', '~> 7.0'
 gem 'mysql2', '~> 0.5.0' # 0.5 and higher are not supported by Rails < 5'
 gem 'uglifier', '>= 1.3.0'
 gem 'bootsnap', '>= 1.0.0', require: false
 gem 'turbolinks', '~> 5'
 gem 'jbuilder', '~> 2.5'
+
+# The original asset pipeline for Rails [https://github.com/rails/sprockets-rails]
+gem "sprockets-rails"
 
 gem 'rake'
 
