@@ -61,7 +61,7 @@ class LocationsController < ApplicationController
     @location = Location.find(params[:id])
 
     respond_to do |format|
-      if @location.update_attributes(location_params)
+      if @location.update(location_params)
         format.html { redirect_to(@location, notice: 'Location was successfully updated.') }
         format.xml  { head :ok }
       else
