@@ -9,8 +9,7 @@ Bundler.require(*Rails.groups)
 module RailsPortal
   class Application < Rails::Application
     # Initialize configuration defaults for originally generated Rails version.
-    config.load_defaults 6.0
-    config.autoloader = :zeitwerk
+    config.load_defaults 7.0
 
     # Configuration for the application, engines, and railties goes here.
     #
