@@ -10,16 +10,15 @@
 #
 # It's strongly recommended that you check this file into your version control system.
 
-ActiveRecord::Schema.define(version: 2023_03_09_210247) do
-
+ActiveRecord::Schema[7.0].define(version: 2023_03_10_214221) do
   create_table "authors", charset: "latin1", force: :cascade do |t|
     t.string "author"
     t.string "author_first"
     t.string "author_last"
     t.string "title"
     t.string "other"
-    t.datetime "created_at"
-    t.datetime "updated_at"
+    t.datetime "created_at", precision: nil
+    t.datetime "updated_at", precision: nil
     t.string "name"
   end
 
@@ -28,8 +27,8 @@ ActiveRecord::Schema.define(version: 2023_03_09_210247) do
     t.integer "author_id"
     t.boolean "translator", default: false
     t.boolean "editor", default: false
-    t.datetime "created_at"
-    t.datetime "updated_at"
+    t.datetime "created_at", precision: nil
+    t.datetime "updated_at", precision: nil
     t.boolean "contributor", default: false
   end
 
@@ -37,8 +36,8 @@ ActiveRecord::Schema.define(version: 2023_03_09_210247) do
     t.string "image_url"
     t.text "bio_text"
     t.integer "author_id"
-    t.datetime "created_at", null: false
-    t.datetime "updated_at", null: false
+    t.datetime "created_at", precision: nil, null: false
+    t.datetime "updated_at", precision: nil, null: false
   end
 
   create_table "books", charset: "latin1", force: :cascade do |t|
@@ -50,8 +49,8 @@ ActiveRecord::Schema.define(version: 2023_03_09_210247) do
     t.string "ISBN"
     t.integer "pages"
     t.integer "original_language"
-    t.datetime "created_at"
-    t.datetime "updated_at"
+    t.datetime "created_at", precision: nil
+    t.datetime "updated_at", precision: nil
     t.boolean "excluded", default: false
     t.string "sortby_title"
     t.integer "genre_id"
@@ -59,21 +58,21 @@ ActiveRecord::Schema.define(version: 2023_03_09_210247) do
 
   create_table "genres", charset: "latin1", force: :cascade do |t|
     t.string "name"
-    t.datetime "created_at"
-    t.datetime "updated_at"
+    t.datetime "created_at", precision: nil
+    t.datetime "updated_at", precision: nil
     t.boolean "readable", default: true
   end
 
   create_table "languages", charset: "latin1", force: :cascade do |t|
     t.string "name"
-    t.datetime "created_at", null: false
-    t.datetime "updated_at", null: false
+    t.datetime "created_at", precision: nil, null: false
+    t.datetime "updated_at", precision: nil, null: false
   end
 
   create_table "locations", charset: "latin1", force: :cascade do |t|
     t.string "name"
-    t.datetime "created_at"
-    t.datetime "updated_at"
+    t.datetime "created_at", precision: nil
+    t.datetime "updated_at", precision: nil
     t.boolean "readable", default: true
   end
 
@@ -81,11 +80,11 @@ ActiveRecord::Schema.define(version: 2023_03_09_210247) do
     t.integer "book_id"
     t.integer "rating"
     t.text "comments"
-    t.datetime "date_started"
-    t.datetime "date_finished"
-    t.datetime "updated_on"
-    t.datetime "created_at"
-    t.datetime "updated_at"
+    t.datetime "date_started", precision: nil
+    t.datetime "date_finished", precision: nil
+    t.datetime "updated_on", precision: nil
+    t.datetime "created_at", precision: nil
+    t.datetime "updated_at", precision: nil
   end
 
 end
